// src/models/enums.tsx
export enum HouseId {
  Empty = '',
  Abhorrers = 'abhorrers',
  Carcass = 'carcass',
  Deadsouls = 'deadsouls',
  Gargamox = 'gargamox',
  Goregrinders = 'goregrinders',
  Igorri = 'igorri',
  Steeplewrack = 'steeplewrack',
  Heretic = 'heretic',
}

export enum UnitTypeId {
  Empty = '',
  Thrall = 'thrall',
  Scion = 'scion',
  Freak = 'freak',
  Horror = 'horror',
  Hunter = 'hunter',
  Tyrant = 'tyrant',
  Necromancer = 'necromancer',
}

export enum UpgradeType {
  Trait = 'trait',
  Ability = 'ability',
  SoulAbility = 'soulAbility',
}